REFACTOR - Pulled WTS game func from util folder
