+++ server/middleware/auth.js
@@ -5,7 +5,6 @@ module.exports = function (req, res, next) {
     const token = req.header('x-auth-token');
     if (!token) return res.status(401).send('Access denied, No token provided...');
 
-
     try {
         const decoded = jwt.verify(token, config.get('jwtPrivateKey'));
         req.user = decoded;
